+++ src/server/player/manager.h
@@ -5,6 +5,9 @@
 
 #include <core/entrypoint.h>
 #include <public/playerslot.h>
+#include <steam/steam_api_common.h>
+#include <steam/isteamuser.h>
+#include <steam/steamclientpublic.h>
 
 class PlayerManager
 {
@@ -29,6 +32,10 @@ public:
     bool ClientConnect(CPlayerSlot slot, const char* pszName, uint64 xuid, const char* pszNetworkID, bool unk1, CBufferString* pRejectReason);
     void OnClientConnected(CPlayerSlot slot, const char* pszName, uint64 xuid, const char* pszNetworkID, const char* pszAddress, bool bFakePlayer);
     void ClientDisconnect(CPlayerSlot slot, ENetworkDisconnectionReason reason, const char* pszName, uint64 xuid, const char* pszNetworkID);
+
+    void SteamAPIServerActivated();
+
+    STEAM_GAMESERVER_CALLBACK_MANUAL(PlayerManager, OnValidateAuthTicket, ValidateAuthTicketResponse_t, m_CallbackValidateAuthTicketResponse);
 };
 
 extern PlayerManager g_playerManager;